only allow creation, update and removal of places for owned apts

+++ src/resolvers/Mutation/place.ts
@@ -2,8 +2,14 @@ import { Context } from "./../../utils";
 import { getUserId } from "../../utils";
 
 export const place = {
-  createPlace(parent, args, ctx: Context) {
-    // const userId = getUserId(ctx);
+
+  async createPlace(parent, args, ctx: Context) {
+    const userId = getUserId(ctx);
+    const aptUser = await ctx.prisma.apartment({ id: args.apartmentId }).user();
+    if (aptUser.id != userId) {
+      throw new Error("Error, tried to add a place to someone else's apartment");
+    };
+
     return ctx.prisma.createPlace({
       name: args.name,
       difficulty: args.difficulty,
@@ -12,8 +18,14 @@ export const place = {
       }
     });
   },
-  updatePlace(parent, args, ctx) {
-    // const userId = getUserId(ctx);
+
+  async updatePlace(parent, args, ctx) {
+    const userId = getUserId(ctx);
+    const aptUser = await ctx.prisma.place({ id: args.id }).apartment().user();
+    if (aptUser.id != userId) {
+      throw new Error("Error, tried to update a place in someone else's apartment");
+    };
+
     const data = { ...args };
     delete data.id;
 
@@ -22,7 +34,15 @@ export const place = {
       where: { id: args.id }
     });
   },
-  deletePlace(parent, args, ctx) {
+
+  async deletePlace(parent, args, ctx) {
+    const userId = getUserId(ctx);
+    const aptUser = await ctx.prisma.place({ id: args.id }).apartment().user();
+    if (aptUser.id != userId) {
+      throw new Error("Error, tried to remove a place from someone else's apartment");
+    };
+
     return ctx.prisma.deletePlace({ id: args.id });
   }
+
 };
